Camera: Always release the extension session after configuration
Since both the session 'onConfigured' callback and the
initialization session callback run within the same Handler
thread, it is possible for the 'onSuccess' callback to fail
to execute if there is another thread the releases the
extension session and quits the thread.
Ensure that the extension session is released also when we
have a valid capture session.

Bug: 270276341
Test: atest -c -d
cts/tests/camera/src/android/hardware/camera2/cts/CameraExtensionSessionTest.java

Change-Id: I03fd9b2859ae8919ca89c823fed973853030bc5f